bc layer for translatable factory interface
