+++ src/Component/spec/Factory/TranslatableFactorySpec.php
@@ -14,11 +14,11 @@ declare(strict_types=1);
 namespace spec\Sylius\Component\Resource\Factory;
 
 use PhpSpec\ObjectBehavior;
-use Sylius\Component\Resource\Exception\UnexpectedTypeException;
-use Sylius\Component\Resource\Model\TranslatableInterface;
 use Sylius\Component\Resource\Translation\Provider\TranslationLocaleProviderInterface;
 use Sylius\Resource\Factory\FactoryInterface;
+use Sylius\Resource\Factory\TranslatableFactory as NewTranslatableFactory;
 use Sylius\Resource\Factory\TranslatableFactoryInterface;
+use Sylius\Component\Resource\Factory\TranslatableFactoryInterface as LegacyTranslatableFactoryInterface;
 
 final class TranslatableFactorySpec extends ObjectBehavior
 {
@@ -32,28 +32,13 @@ final class TranslatableFactorySpec extends ObjectBehavior
         $this->shouldImplement(TranslatableFactoryInterface::class);
     }
 
-    function it_throws_an_exception_if_resource_is_not_translatable(FactoryInterface $factory, \stdClass $resource): void
+    function it_implements_legacy_translatable_factory_interface(): void
     {
-        $factory->createNew()->willReturn($resource);
-
-        $this
-            ->shouldThrow(UnexpectedTypeException::class)
-            ->during('createNew')
-        ;
+        $this->shouldImplement(LegacyTranslatableFactoryInterface::class);
     }
 
-    function it_creates_translatable_and_sets_locales(
-        FactoryInterface $factory,
-        TranslationLocaleProviderInterface $localeProvider,
-        TranslatableInterface $resource,
-    ): void {
-        $localeProvider->getDefaultLocaleCode()->willReturn('pl_PL');
-
-        $factory->createNew()->willReturn($resource);
-
-        $resource->setCurrentLocale('pl_PL')->shouldBeCalled();
-        $resource->setFallbackLocale('pl_PL')->shouldBeCalled();
-
-        $this->createNew()->shouldReturn($resource);
+    function it_should_be_an_alias_of_translatable_factory_interface(): void
+    {
+        $this->shouldBeAnInstanceOf(NewTranslatableFactory::class);
     }
 }
