+++ src/Component/tests/spec/Factory/FactorySpec.php
@@ -11,12 +11,10 @@
 
 declare(strict_types=1);
 
-namespace spec\Sylius\Component\Resource\Factory;
+namespace spec\Sylius\Resource\Factory;
 
 use PhpSpec\ObjectBehavior;
-use Sylius\Resource\Factory\Factory as NewFactory;
 use Sylius\Resource\Factory\FactoryInterface;
-use Sylius\Component\Resource\Factory\FactoryInterface as LegacyFactoryInterface;
 
 final class FactorySpec extends ObjectBehavior
 {
@@ -28,16 +26,6 @@ final class FactorySpec extends ObjectBehavior
     function it_implements_factory_interface(): void
     {
         $this->shouldHaveType(FactoryInterface::class);
-    }
-
-    function it_implements_legacy_factory_interface(): void
-    {
-        $this->shouldHaveType(LegacyFactoryInterface::class);
-    }
-
-    function it_is_an_alias_of_the_factory(): void
-    {
-        $this->shouldBeAnInstanceOf(NewFactory::class);
     }
 
     function it_creates_a_new_instance_of_a_resource(): void
